data_servers/pyproject.toml
[tool.poetry]
name = "data_servers"
version = "0.1.0"
description = ""
authors = ["Andreea Vasiliu <58026816+anndreeaagml@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.10"
uvicorn = "^0.18.3"
fastapi = "^0.85.0"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
